delegates comparison to lambda functions [Java]
COMMENTS:
Code shows how lambda functions may be used to define how to
compare list elements when applying a sorting method.

In this revision the Insertion Sort method has been overloaded so
that it can delegate the comparison of list elements to a
user-defined lambda function. Depending on the comparison function
passed to the Insertion Sort method, it sorts a list of integers in
ascending or descending order.