simplifies swapping on first half of list [Java]
COMMENTS:
Swaps the smaller element with the pivot so that the smaller one
ends up after the pivot element. Then it updates the indexes so
that `i' refers to the element right before the pivot (p - 1) and
`p' refers to the pivot element (as it should). This is simpler
than doing the pair of swaps on the same iteration.