updates median computation [Java]
COMMENTS:
This is a step towards completion of task #32 (median computation).
Uses the FindKthLargest Algorithm to compute the median of a list.
If the list has an odd number of elements, it simply uses the
middle value of the partitioned list; otherwise it averages the
middle element and previous element of the partitioned list to
obtain the median.