validates computation of the median [Java]
COMMENTS:

Changes on the implementation of the partition scheme:
Swaps larger elements with elements from the front of the list
instead of decrementing the index `i' on the first half.
Note that if the algorithm decrements the index `i' when a larger
element is found, it also has to swap the pivot twice on the next
iteration so that the partitioning is preserved. That is, elements
to the right of the pivot are less than the pivot. And that is
important because that is the objective of the left() method: to
move all the elements smaller than the pivot on the firt half after
the pivot element or equivalently to place them in the asymmetric
range [p + 1, m).

The test code validates the adopted partitioning scheme by finding
the median of a set of random numbers. If the partitioning fails
the test throws a runtime exception to inform the user. This is the
final step towards completing task #32