falls back to sequential search [Java]
COMMENTS:
When the user passes a custom comparator to the sorting method
the default comparison scheme used by Binary Search might be
inconsistent with the current ordering. For this reason one has
no choice other than to use sequential searching in those cases,
this is why the sorted state has been set to false. And for good
reason since that has been confirmed by tests (not committed).