pops elements in order to construct a sorted Ry
COMMENTS:
Elements in Py macthing those in Rx are popped in order so that Ry
ends up with the same ordered structure as Py. This is better than
`randomly' looking for them to construct Ry and having to sort Ry
so that it can be used to partition the particles into quadrants.

Note that Rx has been sorted in the default way so that searching
elements in Rx is O(log(N)) because the vector would use Binary
Search. On the other hand, sorting in the average is O(N^2) because
the vector uses Insertion Sort. Thus, the `added complexity' of the
lines of code for the construction of both Ly and Ry is duly
justified.

Ultimately the revised code ends up with the expected quadrants
partitioning.

Comments have been added to the code to explain the handling of the
index used for traversing Py while popping elements from it.