caters empty dataset possibility [Java]
COMMENTS:
If there are no closest-pair candidates near the boundaries it is
possible that the Left subset will have zero elements after the
pruning of elements ends. In that case there's nothing left to do
other than return and signal the caller that there are no better
closest-pair candidates, as done here.