removes elements from the back of vector for speed
COMMENTS:
Removes elements from the back of the Left subset (vector) for
speed. As a result of that whenever an element is popped from the
vector only those which are closest-pair candidates are copied.
In the former case most of the elements had to be copied
(behind scenes) from the onset.

The process of removing particles too far to be closest-pairs in
the Left subset now has to consider all the particles from the back
to the front of the vector so that only the candidates remain. This
is why the for-loop does not break ``early''. Despite this the
process of removing elements from the vector is O(N log(N)) because
of the search that pop does.

Note that the index `idx' needs updating on every pass to consider
all the elements in the vector. It was not a problem for the Right
subset since the process ends the first time it encounters a
closest-pair candidate. Nevertheless, it has been moved from the
body of the nested if-conditional even if that does not affect the
final outcome.

A runtime check has been done to verify that there are few
closest-pair candidates each time the middles are considered.

We can do better by popping elements from the vector by index
instead.