pops vector elements by index for speed [Java]
COMMENTS:
Since the pop-by-index method was not implemented at the time of
development of the code that solves the closest pair problem,
there was no other choice other than to pop elements by their
values. The time complexity of constructing Mx and My subsets
is expected to be linear with respect to the vector size O(N).