improves division of Py subsets into quadrants
COMMENTS:
Improves the time complexity of Dividing Ly (Ry) subsets into
quadrant II, III (I, IV) subsets from O(N^2) to O(N log N).

Note that searching for elements in the Rx subset is O(log N)
because Binary Search is used owing to the ordered structure of
this subset; thus, construction of Ry is O(N log N) because the
process requires a search for the N elements of Py in Rx.

By following a similar train of thought one can deduce that
construction of Ly is also O(N log N). Note that the algorithm
removes elements from the back of the Ly subset vector, thus,
the number of comparisons carried out by pop() behind the scenes
to shift invalidated elements (those after the popped element)
should not be comparable with the vector size N.

On the other hand, removing elements from the front of a vector
(as before) incurs in an overhead comparable to the vector size N,
making the construction of the Ly subset O(N^2).

This summarizes the rationale of this revision.