initial import of the game backend
COMMENTS:
The imported code is based on the Let's Build a Game Tutorial 1.
References shall be added in a future revision.

Defines the basic features (dimensions) of the game window.

Creates the initial game class.

The game executes in a synchronized thread. Implements the start
and stop methods and the game-loop.

The author of the tutorial mentioned that he borrowed the game
loop. He did not explain the implementation of the game-loop,
the comments are based on my interpretation and documentation of
the invoked System methods.