extends the game backend
COMMENTS:
Based on Let's Build a Game Tutorial 2.

Creates the Game Object class to define all the objects of the
game: players, enemies, items, platforms, etc.

Creates Identifiers IDs for game objects (players and enemies)
via Java's enumerated data type `enum'.

Creates an initial player class for demonstration purposes.
The game spawns four ``players'' at random locations, and these
move autonomously to the right with a constant velocity.

Creates a Handler for rendering and updating game objects.

Comments have been added to the game loop to improve the initial
explanation.

Adds a ticks counter to the game loop to determine how many
times game objects need to be updated instead of using the tick
accumulator to do that. I hope this change shows better what is
being done.

Implements the tick method of the Game. Essentially the tick method
delegates its task to the Handler.

Updates the render method of the Game so that it also renders
game objects.