processes input from player
COMMENTS:
Based on Let's Build a Game Tutorial 3.
The demo consists of a single player that can be moved with the
keyboard arrows.

Creates Handler and Game Objects prior to the Window as it should
be. Note that the Window starts the game so the handler needs to
be instantiated first, otherwise we are taking the risk of
stumbling upon a null pointer exception (buggy application).

Creates a KeyInput class to process commands from the player.

Adds a key listener to the game to bind key inputs with player
movements.

Reduces the buffer strategy from four to three to make player
movements more fluent.

Removes the constant velocity components from the player
constructor because these now are set by the keyboard input.