adds `bouncer' enemies and heads-up display
COMMENTS:
Based on Let's Build a Game Tutorial 4.

Creates a heads-up display HUD. As a demo, the health bar
decreases every time until it is depleted.

Creates the Basic Enemy Class `Bouncers'. These enemies move
around with a constant speed, whenever they hit the game walls
they switch the direction of their movement without loosing speed.

Implements a general purpose clamp() method to impose limits on
variables.

Implements Rigid and Periodic Boundaries for the Player.

Improves game fluency by reducing the buffer strategy.
(This is a change I decided to make to improve the fluency of the
game when it is running in my PC.)

Request focus so that the user does not have to click on the game
window to be able to control the player.