adds collisions, trails, and garbage collector
COMMENTS:
Based on Let's Build a Game Tutorial 5.

Disables draining health points from the HUD on each tick since
that was just a demo of the previous tutorial.

Implements a garbage collector to remove objects from the game.
Adds internal garbage state to the GameObject class in order
identify which objects are to be removed.

Drains player health upon collisions with enemies. Implements this
by creating a rectangular `mask' that surrounds both the enemies
and the player. If there's an intersection (or overlap) between
them the code drains the player's health. Intersections are
detected via the intersections() method provided by the Rectangle
class. Adds the getBounds() method to game objects for collision
handling.

Adds IDs for after-image or trail objects.
Implements an after image effect or trail for game objects. This
is done by creating a trail object with the same dimensions and
shape as the game object at its current position x, y. The trail
object has a timespan that's depleted on every time (or tick).
After the timespan reaches a certain value the trail object is
marked as garbage and subsequently removed by the garbage
collector. The timespan is also used to change the alpha composite
value to fade the trail object and thus create an after image
effect.

Adds color, dimension, trailspan, and shape attributes to both the
Player and BasicEnemy classes. Also adds a handler to create after
image effects or trails. Revises the render method to account for
different shapes and colors.

Overloads the clamp() method of the Game class for floats and
doubles.