adds score, level, spawner, projectiles, etc.
COMMENTS:
Based on Let's Build a Game Tutorial 6.

Creates the Fast Enemy class whose movement in the y-direction
surpasses that of the basic enemy, though it moves a bit slower
in the x-direction. Naturally, if the player is hit by this
enemy the damage received is greater.

Implements Spawner utility that spawns enemies once the player
reaches a new level. Leveling up happens every time the player
scores a certain number of points.

Improves the heads-up display HUD by showing the score and level.
Displays the player health as a red bar when low.

Creates Projectile objects by extending the Game Object class.
Adds the shoot method for Game Objects. Enables the player the
ability to shoot at enemies with projectiles. The keyboard
spacebar is used to toggle the firing mode.

Removes enemies hit by projectiles by marking them first as
destroyed objects followed by removal by the garbage collector.
Similarly removes projectiles that hit enemies and those that
reach the game boundaries.

Revises the tick method of the Handler to account for changes
in the size of the linked list. This is achieved by using a while
loop instead of a for-loop. Note that the invoking the tick method
of objects will change the linked-list size because of the creation
of trails (after image effects) and projectiles. I had run into
runtime issues with the for-loop because of the size change of the
linked-list. The implementation with the while-loop fixes the
problem.

The tick method of enemies takes into account the destroyed state
of the object prior to performing updates.

Improve handling of key events by making the player the first
object in the linked-list. By doing this we are certain that the
player will always be at the front of the linked-list regardless
of the insertion or removal of other game objects. In my opinion
it's a better design to reserve the first couple of elements of
the linked-list for special objects.