creates smarter enemy
COMMENTS:
Based on Let's Build a Game Tutorial 7.

Creates a new type of enemy that follows the player. This is
implemented by applying vector algebra: the direction of the
velocity of the enemy is somewhat aligned with the relative
position vector. In contrast to the approach used by the author
of the tutorial I decided to stick with the integral type for
both the position and velocity vectors. What I did to improve
the ability of the enemy to follow the player was to add a bias
component to the velocity of the enemy.

Defines width and height as attributes of Game Objects.
Since we need the dimensions of the player and smart enemy to
correct their position vectors to their centers these attributes
have been added to the Game Object Class for ease of access via
suitable getters.

Since the only game objects that have dimensions are the enemies,
the trails, the projectiles, and the player we are using their IDs
to define their sizes. This might change in a future revision as
new objects are added. For the trail and the projectiles though we
are simply using the new GameObject constructor that takes as
arguments the width and height.

To simulate the death of the player we simply ignore the key inputs
from the user when the player health gets depleted totally. Later
on this can be improved to show a Game Over text or something like
that.

For the projectile we are making its x-axis velocity equal to that
of the player just to make it a bit more realistic.

The spawner now spawns smart enemies on level ups as one would
expect from the context of this commit.

If the user presses the R key the player gets it health
Replenished. This is useful if I want to recover control of the
player for further testing. This feature might be removed in a
future revision.