spawns bosses at level 10 and 20
COMMENTS:
Creates a Boss Enemy Class derived from the Game Object Class.
The Boss can be regarded as a dropship that spawns enemies instead
of a larger enemy that shoots at the player. If the player
collides with the enemy boss all its health gets depleted.
Even though projectiles get destroyed upon impact with the Boss,
these do not damage the Boss at all. We do this to simulate an
indestructible boss. We use a counter to time the boss fight
duration to five levels.

The spawning system has been modified so that no more enemies
are spawned after level 20. Note that this might change in a
future revision. The number of basic enemies spawned now depends
on the level. After reaching level 15 we spawn a group of smart
enemies instead of just one.

The placement of the projectiles that the player shoots have been
corrected by the projectile width to improve alignment.

The game window has been increased by 50% while leaving the object
sizes constant.