adds menu and help screens to Swarm Game
COMMENTS:
Based on Let's Build a Game Tutorial 9.

Creates basic menu and help screens with buttons that can be
clicked by the user with the mouse.

If the user clicks on the Play button from the Menu the game is
initialized --- creates the player and a group of basic enemies.
And from there the game proceeds as usual.

Adds a dispose() method to dispose of the game window when the
user clicks on the quit button.

The game tick() and render() methods now depend on the game state
so that the menu, help, or the game is rendered.

Portion of the code has been reorganized and some comments have
been modified to provide a more accurate description.