relocates methods in source file [clang]
COMMENTS:
Moves the constructor close to the end of the source file.

Rationale:
To make the vector OOP-like the methods need to be defined before
the constructor. This is because the constructor binds function
pointers to the methods (of the vector class) and to be able to do
that the methods must be defined first.