performs matrix operations concurrently [clang]
COMMENTS:
Uses POSIX threads to parallelize vector sum and matrix
multiplication. The outermost loop of the matrix multiplication
routine is evenly distributed among threads, and it also avails
itself of the fact that the matrices are symmetric to maximize
the use of the cache of the CPU.

Valgrind reports that there are *no* memory leaks.