adds critical section example [clang]
COMMENTS:
Computes the vector scalar product with multiple threads.

The code is based on the POSIX threads tutorial of the
Lawrence Livermore National Laboratory LLNL:

https://hpc-tutorials.llnl.gov/posix/example_using_mutexes

In contrast to the original work default thread attributes have
been used for numeric computation. Also, the shared resource is
passed as an argument to the method instead of using the global
scope.

Each thread operates on a contiguous slice and writes the
intermediate result (its respective sum of squares) to the
shared resource (scalar quantity). The writing of intermediate
results is carried out within a critical section. This means that
only one thread can write to the shared resource at a time;
otherwise, a data race condition would exist and the numeric
result (the scalar product) would be unreliable.

Valgrind reports no memory leaks.