implements copy constructor for vector [Java]
COMMENTS:
Implements a copy constructor for the vector class.

The copy constructor is used in the closest pair problem to protect
the original data (particle coordinates). The partitioning is done
on the copy, for the divide and conquer algorithm should not alter
the data in any way.

Implements a method to return a copy of the data stored at a given
location.

Implements a method to partition the data into two parts. In the
context of the closest pair problem, we use this method to create
the left and right partitions.

Implements a print method to display the data on the console.