plots the full range of the elapsed-time [python]
COMMENTS:
Plots the full range of the elapsed-time data.

This is the version used to generate the plot for the presentation
of the closest-pair problem.

Both the elapsed time and the number of distance computations
exhibit a linear dependence with the input size.

The time-complexity study was repeated because the HPCf node
executing the code was running with more jobs that it should
handle.