shows that the runtime is O(N log N) [Python]
COMMENTS:
Even though the average number of distance computations grows
linearly with the system size N, the runtime exhibits a O(N log N)
dependence. This means that the cost of combination plus division
grows linearly. It is possible that the time spent on dividing the
vector into two partitions comes into play for larger N.