uses a singly linked-list to find the closest pair
COMMENTS:
uses a singly linked-list to find the closest pair [Java]

Uses a singly linked-list to store the 2D Cartesian coordinates of
an ensemble of particles. Both the implementation of the Brute
Force and Divide and Conquer Algorithms get access to the particle
coordinates through the methods of the singly linked-list class.

The time complexity plots shows the effect of the linear search of
the singly linked-list on the runtime performance of the closest
pair seeking algorithms.