uses a tree to find the closest pair [Java]
COMMENTS:
The Tree Class is very similar to that of the Vector Class with
the exception that we keep the elements sorted while inserting
them.

The Ensemble and Time Classes are essentially the same as those
used with a Vector other than using a different container.

No differences in the time complexity have been found on the
closest pair seeking algorihtms, as expected, for the underlying
data structure is an array for both containers.