makes the default pair object more useful [Java]
COMMENTS:
The default constructor of the Pair class was introduced at first
just to get rid of the complain issued by the Java Compiler about
uninitialization.

By making the default pair object to be comprised by infinitely
separated points we can now give the default constructor a better
use, as it will be shown in the next commits where the Brute Force
algorithm will be revised to take into account this change.