defines method for spawning points in a 2D domain
COMMENTS:
Instead of overloading the existing create() method, I have opted
to define a new create2D() method to make the code easier to read
and understand. The create() method creates an ensemble of points
to solve the 1D Closest Pair problem, and the create2D() method
creates an ensemble of points to solve the 2D counterpart.

In a future commit a create3D() method will be added to tackle the
3D Closest Pair problem.

No other changes have been done that modify what the code does at
the fundamental level.