makes the Point class extendable [Java]
COMMENTS:
By doing this we can define the 3D Point class as a subclass of the
Point class. The components of the Point class are still final and
private so that the Point3D class cannot alter their values.

Bounds the distance computing method to objects of the Point class.
This has done so that the derived class Point3D can override it
accordingly.

Makes available the static method compare(double, double) to the
Point3D class to avoid needless repetition.

The Ensemble class has been modified to account for these changes.

The code passes the existing tests successfully.