solves the 3D Closest Pair problem [Java]
COMMENTS:
The implementation of the 3D Divide And Conquer Algorithm exhibits
the expected theoretical time complexity of a recursive algorithm
whose division plus combination steps are O(N log N).

The implementation of the 3D Divide And Conquer Algorithm is
analogous to that of its 2D counterpart, as it can be seen from
the source code.

The implementation passes the tests successfully, meaning that the
Brute Force and Divide And Conquer Algorithms find the same closest
pair.