uses a double to store the ensemble size [Java]
COMMENTS:
Uses a double to store the ensemble size. This is important when
computing the limits for the x, y, z coordinates of the points.
Note that for large ensembles the computation of the limit could
overflow unless Java casts the integral size into a double before
computing the product. I prefer to make the conversion myself to
err on the safer side.