solves the 1D closest pair problem [cxx]
COMMENTS:
Passes the implementation check test, that is, the closest pair
found by the 1D Divide And Conquer Algorithm matches the closest
pair found by the Brute Force Algorithm.

The 1D Divide And Conquer Algorithm exhibits the expected theoretical
linear time complexity.

The implementation of the console application is analogous to its
Java counterpart in the sense that objects are created and destroyed
dynamically. Java does a better job at managing memory this way and
this could be the possible reason for the unsatisfactory relative
runtime performance. There could be other factors but further testing
is needed. Nevertheless, the average runtimes of the implemented
1D Divide and Conquer Algorithm in C++ and Java are comparable.

It would be interesting to see if the implementation in C++ improves
by removing the burden of managing dynamically allocated objects.

Valgrind reports that the console application does not incur on
memory leaks.