minimizes dynamically allocated objects [cxx]
COMMENTS:
Minimizes the use of dynamically allocated objects in the rest
of the source files.

To achieve this a copy method was implemented for the Point class.
Had to change access to the Point components from private to
protected to implement the copy method for 3D Point objects with
ease. There was really no reason to make them private in the first
place though other than be overly cautious.

The overall runtime performance of the console application is
comparable to that of the previous commit, this confirms that the
bottleneck of the application is the Brute Force Algorithm. Thus,
other changes to the source code will not improve significantly the
overall runtime.