solves the 1D Closest Pair Problem [FORTRAN]
COMMENTS:

Passes the implementation check test, that is, the closest pair
found by the 1D Divide And Conquer Algorithm matches the closest
pair found by the Brute Force Algorithm for all the considered cases.

The 1D Divide And Conquer Algorithm exhibits the expected theoretical
linear time complexity.

The overall runtime of the application cannot be fairly compared
against its Java and C++ counterparts because the sorting algorithm
is quadratic; nevertheless, I am reporting it here:

[@hostname fortran]$ time ./test-closest-pair

real	0m57.033s
user	0m57.019s
sys	0m0.004s

My implementations of the 1D Divide and Conquer Algorithm in Java
and C++ are at least three times faster (for an ensemble of 2048
points). I can only guess that the (de)construction of objects
(even if limited in a number) significantly retards the execution
speed of the implementation in FORTRAN. It is also possible that
some compiler optimizations were missed owing to the OOP features
of the code. This could be the reason that experts in scientific
computing usually recommend working with arrays of intrinsic (or
primitive) types.

It would be interesting to repeat the experiment with binaries
produced by other compilers other than GCC.

Valgrind reports that the console application does not incur on
memory leaks.