removes object (de)construction overhead [FORTRAN]
COMMENTS:
Removes object (de)construction overhead from the implementation of
the 1D Divide And Conquer Algorithm that solves the closest pair
problem.

As a result of this, the implementation in FORTRAN is at least 1.5
times faster than the implementation in Java and at least 2 times
faster than the implementation in C++ (for ensembles of size 2048).
Confirming that the overhead of (de)constructing objects dynamically
was significant, as noted in the previous commit.

For academic reasons both versions of Divide And Conquer Algorithm
have been kept.

Nevertheless, the overall runtime of the application in FORTRAN has
not changed significantly, for the sorting algorithm (which runs on
quadratic time) determines the overall application runtime:

[@hostname fortran]$ time ./test-closest-pair

real	0m53.971s
user	0m53.871s
sys	0m0.044s

Valgrind reports that memory leaks are not possible.