revises comments in combine() methods [Java]
COMMENTS:
Revises comments in combine() methods.

The combine() method calculates the distance of closest pair
candidates at partition interfaces (between the left and the right
partition). The method picks points in the left parition that are
closest to the interface (that is, rightmost points in the left
partition) and picks points in the right partition that closest to
the interface (that is, leftmost points in the right partition) to
check if these are closer than the closest pair.

The revision attempts to express this more clearly in specific
lines of code.