solves 1D Closest Pair Problem with List<T> [Java]
COMMENTS:
Solves the 1D Closest Pair Problem with a user-defined, generic,
iterable, singly, linked-list List<T>. The implementation passes the
usual tests: both the Brute Force Algorithm and the Divide And
Conquer Algorithms report the same closest pair for all the
cases considered.

The implementation of the divide and conquer algorithm exhibits,
in the average sense, a linear time complexity; however, the
implementation that uses the generic Vector<T> is about two times
faster (checked for ensembles comprising 2048 points).