bruteForceMethod -> timeBruteForce for readability
COMMENTS:
Removes the string `Method' from the name of `Method' ending
methods, functions, subroutines, etc.

These methods are used for timing the implementation of the
Brute Force Algorithm and for this reason the string `time' has
been added at the front of their names to make them easier to
read.

Simirly, other methods used for the same purpose have been renamed
accordingly.

This applies to the implementation of the closest pair finding
algorithms in Java, C++, and FORTRAN.