solves the 1D Closest Pair Problem [Python]
COMMENTS:

Passes the implementation check test, that is, the closest pair
found by the 1D Divide And Conquer Algorithm matches the closest
pair found by the Brute Force Algorithm for all the considered cases.

The 1D Divide And Conquer Algorithm exhibits the expected theoretical
linear time complexity.

The implementation in Python of the Divide And Conquer Algorithm is
over a 100 times slower than its Java counterpart. Again, the
dynamic allocation of objects could be a significant overhead, as
it has been seen in the (earlier versions of the) C++ and FORTRAN
counterparts.

OPTIMIZATIONS:

The code can be optimized by solving the problem entirely with
numpy arrays. A Just-in-Time JIT Python Compiler can be used to
obtain a code that performs closer to the optimal machine code.