revises velocity update smart-enemy [Java]
COMMENTS:
Note that if the distance between the player and the smart-enemy
is not too great the velocity components get overwritten and then
the bias is added.

If they are too close the velocity components are set equal to the
bias.

This is still what the code does but to achieve the same effect
we had to zero the velocity components beforehand.

I have checked that the dynamics of smart enemies remains the same
during game play (as expected).