adds implementation of merge sort [clang]
COMMENTS:
Adds possible implementation of merge sort

The time complexity analysis shows that the implementation of the
merge sort algorithm is O(N * log N), where N is the size of the
array to be sorted.

Adds a sane check in create() to make sure that the array to be
sorted have an even number of elements, as required by merge sort.

Two versions of combine() have been added for academic purposes.

valgrind reports no memory issues