improves interface by adding particle type [clang]
COMMENTS:
By adding the particle type the implementation details (or the
internals) are better hidden from the user; as a consquence of this
the interface looks cleaner.

The users need only to set how many particles (or points) do they
want to run the implementation of the closest pair finding algorithm.

The time complexity of sort() is still O(N log N) as expected.

Valgrind reports no memory issues