finds the closest pair recursively [clang]
COMMENTS:
Finds the closest pair recursively via a divide and conquer scheme

The implemenetation of the divide and conquer algorithm recurse()
passes initial tests; that is, recurse() finds the same closest
pair as the bruteForce() as it should because there's only one
(there are no other pairs having the same separation distance)

The only issue is that recurse() fails to determine the IDs of
the particles that comprise the closest pair. A possible solution
to this issue is to have sort() operate on an array of IDs so
that recurse() can fetch the IDs with ease. Note that the space
complexity is still O(N) where N is the number of particles.
Moreover, it is common for numerical codes to allocate space
for an array of IDs; therefore, this approach is good enough.
And the time complexity of sort() will still be O(N log N); we
can also check that.

A good reason to commit this is that it is a good intermediate
point between the initial and final implementations.

valgrind reports no memory issues.