fixes closest-pair IDs issue [clang]
COMMENTS:
recurse() now properly finds both the distance and the IDs of the
particles that comprise the closest pair.

test-recurse() shows that recurse() obtains the same result as
bruteForce().

code passes existing tests as well

fixes minor issues (compiler warnings and memory deallocations)

valgrind reports no memory leaks

FINAL NOTES:
Had to carefully add the IDs array to the particles; for example,
the search() method assumes that there's a temporary `xtmp' at a
distance 2 * N from `x', where `N' is the number of particles. So
placing the start of the IDs array there would have been a mistake.

It is important to mention that recurse() modifies the `particles'
data owing to the sorting and so one has to be careful when reusing
`particles'. This is because the particle IDs are not in their
original positions and so it would seem that the bruteForce() method
finds a different pair because the loop indexes `i' and `j' no longer
match the IDs (as they should). This is the reason why we use iota()
in initialize() to make sure that the IDs are in the expected order
so that bruteForce() does not need to care (and shouldn't care) about
the IDs array and simply use the loop indexes. In fact, it is the
right thing to make sure that we set all the properties of
`particles' in initialize() since it is an initializer and so it
should do that regardless.