makes pair and point types more OOP-like [clang]
COMMENTS:
Adopts conventions used by python developers to express OOP code;
such as "private" attributes are denoted by a leading underscore,
like `_x' and `_y', to indicate that these are implementation
dependent and may change in the future and so no users should
write code that relies on them.

We have also defined types for these private attributes just for
the sake of making them stand out.

Adds methods for these types which shall be used for solving the
closest pair problem.

adds `static' for implementation functions both to hide them and
to reuse some of their names in other source files