adds optimal solution [clang] [OOP] [Python]
COMMENTS:
Adds optimal implementation (in terms of number of operations)
of the Divide And Conquer algorithm that finds the closest pair.

This implementation partitions the points in O(N) operations at the
expense of allocating the partitions on the heap; despite of these
allocations the space complexity is O(N).

The expected theoretical overall time complexity is O(N log N).

code passes all tests

valgrind reports no memory issues